requirements.txt
python-telegram-bot==13.15
psycopg2-binary==2.9.9
requests==2.32.3
SQLAlchemy==2.0.32
alembic==1.13.2


